jax_face_processing.py
import jax
import numpy as np
from PIL import Image

class JAXFaceProcessor:
    def __init__(self, use_gpu=True):
        self.use_gpu = use_gpu
        self.device = 'gpu' if use_gpu else 'cpu'
        print(f"Initializing JAX processor...")
        print(f"JAX Face Processor using device: {self.device}")
        # Placeholder for model loading
        self.face_detector = None
        self.face_recognizer = None
        self.whitelist_faces = []

    def detect_and_blur_faces(self, frame):
        """
        Detects and blurs faces in a given frame.

        Args:
            frame: The input frame in which to detect and blur faces.

        Returns:
            The frame with detected faces blurred.
        """
        # Placeholder implementation
        print("JAX face detection and blurring is not yet implemented.")
        return frame

    def update_whitelist(self, image_path):
        """
        Updates the whitelist of faces that should not be blurred.

        Args:
            image_path: The path to the image containing the face to be whitelisted.
        """
        # Placeholder implementation
        print(f"JAX whitelist update is not yet implemented for {image_path}.")
        return False
