requirements.txt
PyQt5
opencv-python
numpy
Pillow
torch
torchvision
facenet-pytorch